utils/deploy-yatank-env/common_requirements.txt
backports.ssl-match-hostname==3.5.0.1
certifi==2015.11.20.1
docutils==0.12
lockfile==0.12.2
paramiko==1.17.0
pip==8.1.2
progressbar==2.3 
psutil==4.3.0
pyjade==3.0.0
python-daemon==2.1.1
pytz==2016.4
pyyaml==3.11
requests==2.10.0
setuptools==23.0.0
simplejson==3.8.2
six==1.10.0
slumber==0.7.1
sqlalchemy==1.0.11
termcolor==1.1.0
tornado==4.0
validators==0.10.1
wheel==0.26.0
winpdb==1.3.6
xmltodict==0.10.1
zmq==0.0.0
GitPython==2.0.6
Flask==0.12
supervisor==3.3.1
